Implemented disambiguation of the CMD elements by looking at ConceptLinks in the context. For example:
  <facetConcept name="language">
    <concept>http://www.isocat.org/datcat/DC-2482</concept> <!-- language code -->
    <concept>http://www.isocat.org/datcat/DC-2484</concept> <!-- language name -->
    <acceptableContext includeEmpty="true"/>
    <rejectableContext incudeAny="true">
      <concept>http://www.isocat.org/datcat/DC-4146</concept>
    </rejectableContext>
    <pattern>/c:CMD/c:Components/c:OLAC-DcmiTerms/c:language/@olac-language</pattern>
    <pattern>/c:CMD/c:Components/c:OLAC-DcmiTerms/c:subject/@olac-language</pattern>
    <pattern>/c:CMD/c:Components/c:LrtInventoryResource/c:LrtCommon/c:Languages/c:ISO639/c:iso-639-3-code/text()</pattern>
  </facetConcept>

The acceptable and rejectable context basically work as a white- and blacklist of the first concept found while
walking up from the CMD element to the root. The acceptable and rejectable context can also specify if the include
any context, e.g., in the example basically any context concept encountered will lead to rejection of the path,
and/or the case where no context concept was found, e.g., in the example all of these paths will be accepted. In
general the order is checking what is acceptable and then falling through to checking rejectable. If any context
is allowed in the whitelist the blacklist is first checked. For acceptabe context the default value for
includeEmpty is true and for includeAny is false. For rejectable context its the other way around. When loading
the facet concept mapping its checked if the acceptable and rejectable context of a facet don't both include any
or empty context.

M    java/eu/clarin/cmdi/vlo/importer/FacetConceptMapping.java
- added support for rejectableContext and acceptableContext
- added check() to see if there are contradictory contexts with regard to include empty and any

M    java/eu/clarin/cmdi/vlo/importer/VLOMarshaller.java
- call check() when the facet concept mapping is loaded

M    java/eu/clarin/cmdi/vlo/importer/FacetMappingFactory.java
- if a facet has an acceptable and/or rejectable context this is now checked to determine if the
  path is accepted or not