Issue47 (#90)
This adds processing of the @cmd:ValueConceptLink to the VLO importer. This depends on knowing the vocabulary associated with an element. To do so the XPath and the Vocabulary info found during analyzing the XSD are now stored in the new Pattern and Vocabulary classes.

Added conceptRegistryUrl and vocabularyRegistryUrl to the VloConfig. If the @cmd:ValueConceptLink contains CCR_ do lookup in the CCR, otherwise try CLAVAS.

Implements part of #47